+++ Public/Azure/Request-AzureAudioTranscription.ps1
@@ -50,90 +50,25 @@ function Request-AzureAudioTranscription {
     )
 
     begin {
-        # Initialize API Key
-        [securestring]$SecureToken = Initialize-APIKey -ApiKey $ApiKey
+        # Construct parameters
+        $Parameters = $PSBoundParameters
+        $Parameters.ApiType = [OpenAIApiType]::Azure
+        $Parameters.AuthType = $AuthType
+        $Parameters.Model = $Deployment
+        $null = $Parameters.Remove('Deployment')
 
-        # Initialize API Base
-        $ApiBase = Initialize-AzureAPIBase -ApiBase $ApiBase
-
-        # Get API endpoint
-        $OpenAIParameter = Get-AzureOpenAIAPIEndpoint -EndpointName 'Audio.Transcription' -Engine $Deployment -ApiBase $ApiBase -ApiVersion $ApiVersion
-
-        # Convert language name to ISO-639-1 format (if we can)
-        if ($PSCmdlet.ParameterSetName -eq 'Language' -and $PSBoundParameters.ContainsKey('Language')) {
-            $ci = Get-CultureInfo -LanguageName $Language -ErrorAction Ignore
-            if ($ci -is [cultureinfo]) {
-                $Language = $ci.TwoLetterISOLanguageName
-            }
-        }
-        elseif ($PSBoundParameters.ContainsKey('LiteralLanguage')) {
-            $Language = $LiteralLanguage
-        }
+        # Invoke base function
+        $steppablePipeline = {
+            Request-AudioTranscription @Parameters
+        }.GetSteppablePipeline($myInvocation.CommandOrigin)
+        $steppablePipeline.Begin($PSCmdlet)
     }
 
     process {
-        $FileInfo = (Get-Item -LiteralPath $File)
-        # (Only PS6+)
-        # If the filename contains non-ASCII characters,
-        # the OpenAI API cannot recognize the file format correctly and returns an error.
-        # As a workaround, copy the file to a temporary file and send it.
-        # We need to find a better way.
-        $IsTempFileCreated = $false
-        if ($PSVersionTable.PSVersion.Major -ge 6) {
-            if ($FileInfo.Name -match '[^\u0000-\u007F]') {
-                Write-Warning 'File name contains non-ASCII characters. It is strongly recommended that file name only contains ASCII characters.'
-                $FileInfo = Copy-TempFile -SourceFile $FileInfo -ErrorAction Stop
-                $IsTempFileCreated = $true
-            }
-        }
-
-        #region Construct parameters for API request
-        $PostBody = [System.Collections.Specialized.OrderedDictionary]::new()
-        # No need for Azure
-        # $PostBody.model = $Model
-        $PostBody.file = $FileInfo
-        if ($Format) {
-            $PostBody.response_format = $Format
-        }
-        if ($PSBoundParameters.ContainsKey('Prompt')) {
-            $PostBody.prompt = $Prompt
-        }
-        if ($PSBoundParameters.ContainsKey('Temperature')) {
-            $PostBody.temperature = $Temperature
-        }
-        if ($Language) {
-            $PostBody.language = $Language
-        }
-
-        #region Send API Request
-        try {
-            $Response = Invoke-OpenAIAPIRequest `
-                -Method $OpenAIParameter.Method `
-                -Uri $OpenAIParameter.Uri `
-                -ContentType $OpenAIParameter.ContentType `
-                -TimeoutSec $TimeoutSec `
-                -MaxRetryCount $MaxRetryCount `
-                -ApiKey $SecureToken `
-                -AuthType $AuthType `
-                -Body $PostBody
-        }
-        finally {
-            if ($IsTempFileCreated -and (Test-Path $FileInfo -PathType Leaf)) {
-                Remove-Item $FileInfo -Force -ErrorAction SilentlyContinue
-            }
-        }
-        # error check
-        if ($null -eq $Response) {
-            return
-        }
-        #endregion
-
-        #region Output
-        Write-Output $Response
-        #endregion
+        $steppablePipeline.Process($PSItem)
     }
 
     end {
-
+        $steppablePipeline.End()
     }
 }