+++ Public/Azure/Request-AzureTextCompletion.ps1
@@ -86,167 +86,25 @@ function Request-AzureTextCompletion {
     )
 
     begin {
-        # Initialize API Key
-        [securestring]$SecureToken = Initialize-APIKey -ApiKey $ApiKey
+        # Construct parameters
+        $Parameters = $PSBoundParameters
+        $Parameters.ApiType = [OpenAIApiType]::Azure
+        $Parameters.AuthType = $AuthType
+        $Parameters.Model = $Deployment
+        $null = $Parameters.Remove('Deployment')
 
-        # Initialize API Base
-        $ApiBase = Initialize-AzureAPIBase -ApiBase $ApiBase
-
-        # Get API endpoint
-        $OpenAIParameter = Get-AzureOpenAIAPIEndpoint -EndpointName 'Text.Completion' -Engine $Deployment -ApiBase $ApiBase -ApiVersion $ApiVersion
-
-        # Temporal model name
-        $Model = 'text-davinci-003'
+        # Invoke base function
+        $steppablePipeline = {
+            Request-TextCompletion @Parameters
+        }.GetSteppablePipeline($myInvocation.CommandOrigin)
+        $steppablePipeline.Begin($PSCmdlet)
     }
 
     process {
-        # Prompt is required
-        if ($null -eq $Prompt -or $Prompt.Count -eq 0) {
-            Write-Error -Exception ([System.ArgumentException]::new('"Prompt" property is required.'))
-            return
-        }
-
-        # When BestOf used with NumberOfAnswers, BestOf must be greater than NumberOfAnswers
-        if ($PSBoundParameters.ContainsKey('BestOf') -and $PSBoundParameters.ContainsKey('NumberOfAnswers')) {
-            if (-not ($BestOf -gt $NumberOfAnswers)) {
-                Write-Error -Exception ([System.ArgumentException]::new('BestOf must be greater than NumberOfAnswers'))
-                return
-            }
-        }
-
-        #region Construct parameters for API request
-        $PostBody = [System.Collections.Specialized.OrderedDictionary]::new()
-        # No need for Azure
-        # $PostBody.model = $Model
-        if ($PSBoundParameters.ContainsKey('Prompt')) {
-            if ($Prompt.Count -eq 1) {
-                $PostBody.prompt = [string](@($Prompt)[0])
-            }
-            else {
-                $PostBody.prompt = $Prompt
-            }
-        }
-        if ($PSBoundParameters.ContainsKey('Temperature')) {
-            $PostBody.temperature = $Temperature
-        }
-        if ($PSBoundParameters.ContainsKey('TopP')) {
-            $PostBody.top_p = $TopP
-        }
-        if ($PSBoundParameters.ContainsKey('NumberOfAnswers')) {
-            $PostBody.n = $NumberOfAnswers
-        }
-        if ($PSBoundParameters.ContainsKey('StopSequence')) {
-            $PostBody.stop = $StopSequence
-        }
-        if ($MaxTokens -gt 0) {
-            $PostBody.max_tokens = $MaxTokens
-        }
-        if ($PSBoundParameters.ContainsKey('PresencePenalty')) {
-            $PostBody.presence_penalty = $PresencePenalty
-        }
-        if ($PSBoundParameters.ContainsKey('FrequencyPenalty')) {
-            $PostBody.frequency_penalty = $FrequencyPenalty
-        }
-        if ($PSBoundParameters.ContainsKey('LogitBias')) {
-            $PostBody.logit_bias = Convert-LogitBiasDictionary -InputObject $LogitBias -Model $Model
-        }
-        if ($PSBoundParameters.ContainsKey('User')) {
-            $PostBody.user = $User
-        }
-        if ($PSBoundParameters.ContainsKey('Echo')) {
-            $PostBody.echo = $Echo
-        }
-        if ($PSBoundParameters.ContainsKey('BestOf')) {
-            $PostBody.best_of = $BestOf
-        }
-        if ($Stream) {
-            $PostBody.stream = [bool]$Stream
-            # When using the Stream option, limit NumberOfAnswers to 1 to optimize output. (this limit may be relaxed in the future)
-            $PostBody.n = 1
-        }
-        #endregion
-
-        #region Send API Request (Stream)
-        if ($Stream) {
-            # Stream output
-            Invoke-OpenAIAPIRequest `
-                -Method $OpenAIParameter.Method `
-                -Uri $OpenAIParameter.Uri `
-                -ContentType $OpenAIParameter.ContentType `
-                -TimeoutSec $TimeoutSec `
-                -MaxRetryCount $MaxRetryCount `
-                -ApiKey $SecureToken `
-                -AuthType $AuthType `
-                -Body $PostBody `
-                -Stream $Stream |`
-                Where-Object {
-                -not [string]::IsNullOrEmpty($_)
-            } | ForEach-Object {
-                try {
-                    $_ | ConvertFrom-Json -ErrorAction Stop
-                }
-                catch {
-                    Write-Error -Exception $_.Exception
-                }
-            } | Where-Object {
-                $null -ne $_.choices -and $_.choices[0].text -is [string]
-            } | ForEach-Object {
-                # Writes content to both the Information stream(6>) and the Standard output stream(1>).
-                $InfoMsg = [System.Management.Automation.HostInformationMessage]::new()
-                $InfoMsg.Message = $_.choices[0].text
-                $InfoMsg.NoNewLine = $true
-                Write-Information $InfoMsg
-                Write-Output $InfoMsg.Message
-            }
-            return
-        }
-        #endregion
-
-        #region Send API Request (No Stream)
-        else {
-            $Response = Invoke-OpenAIAPIRequest `
-                -Method $OpenAIParameter.Method `
-                -Uri $OpenAIParameter.Uri `
-                -ContentType $OpenAIParameter.ContentType `
-                -TimeoutSec $TimeoutSec `
-                -MaxRetryCount $MaxRetryCount `
-                -ApiKey $SecureToken `
-                -AuthType $AuthType `
-                -Body $PostBody
-
-            # error check
-            if ($null -eq $Response) {
-                return
-            }
-        }
-        #endregion
-
-        #region Parse response object
-        try {
-            $Response = $Response | ConvertFrom-Json -ErrorAction Stop
-        }
-        catch {
-            Write-Error -Exception $_.Exception
-        }
-        if ($null -ne $Response.choices) {
-            $ResponseContent = $Response.choices
-        }
-        #endregion
-
-        #region Output
-        # Add custom type name and properties to output object.
-        $Response.PSObject.TypeNames.Insert(0, 'PSOpenAI.Text.Completion')
-        if ($unixtime = $Response.created -as [long]) {
-            # convert unixtime to [DateTime] for read suitable
-            $Response | Add-Member -MemberType NoteProperty -Name 'created' -Value ([System.DateTimeOffset]::FromUnixTimeSeconds($unixtime).LocalDateTime) -Force
-        }
-        $Response | Add-Member -MemberType NoteProperty -Name 'Prompt' -Value $Prompt
-        $Response | Add-Member -MemberType NoteProperty -Name 'Answer' -Value ([string[]]$ResponseContent.text)
-        Write-Output $Response
-        #endregion
+        $steppablePipeline.Process($PSItem)
     }
 
     end {
-
+        $steppablePipeline.End()
     }
 }