+++ Public/Request-ImageEdit.ps1
@@ -38,12 +38,24 @@ function Request-ImageEdit {
         [Parameter()]
         [int]$TimeoutSec = 0,
 
+        # [Parameter(DontShow = $true)]
+        # [OpenAIApiType]$ApiType = [OpenAIApiType]::OpenAI,
+
+        [Parameter()]
+        [System.Uri]$ApiBase,
+
+        # [Parameter(DontShow = $true)]
+        # [string]$ApiVersion,
+
+        # [Parameter(DontShow = $true)]
+        # [string]$AuthType = 'openai',
+
+
         [Parameter()]
         [ValidateRange(0, 100)]
         [int]$MaxRetryCount = 0,
 
         [Parameter()]
-        [Alias('Token')]  #for backword compatibility
         [securestring][SecureStringTransformation()]$ApiKey,
 
         [Parameter()]
@@ -55,11 +67,14 @@ function Request-ImageEdit {
         # Initialize API Key
         [securestring]$SecureToken = Initialize-APIKey -ApiKey $ApiKey
 
+        # Initialize API Base
+        $ApiBase = Initialize-APIBase -ApiBase $ApiBase -ApiType [OpenAIApiType]::OpenAI
+
         # Initialize Organization ID
         $Organization = Initialize-OrganizationID -OrgId $Organization
 
         # Get API endpoint
-        $OpenAIParameter = Get-OpenAIAPIEndpoint -EndpointName 'Image.Edit'
+        $OpenAIParameter = Get-OpenAIAPIEndpoint -EndpointName 'Image.Edit' -ApiBase $ApiBase
     }
 
     process {