+++ Public/Request-Moderation.ps1
@@ -20,11 +20,13 @@ function Request-Moderation {
         [int]$TimeoutSec = 0,
 
         [Parameter()]
+        [System.Uri]$ApiBase,
+
+        [Parameter()]
         [ValidateRange(0, 100)]
         [int]$MaxRetryCount = 0,
 
         [Parameter()]
-        [Alias('Token')]  #for backword compatibility
         [securestring][SecureStringTransformation()]$ApiKey,
 
         [Parameter()]
@@ -36,11 +38,14 @@ function Request-Moderation {
         # Initialize API Key
         [securestring]$SecureToken = Initialize-APIKey -ApiKey $ApiKey
 
+        # Initialize API Base
+        $ApiBase = Initialize-APIBase -ApiBase $ApiBase -ApiType [OpenAIApiType]::OpenAI
+
         # Initialize Organization ID
         $Organization = Initialize-OrganizationID -OrgId $Organization
 
         # Get API endpoint
-        $OpenAIParameter = Get-OpenAIAPIEndpoint -EndpointName 'Moderation'
+        $OpenAIParameter = Get-OpenAIAPIEndpoint -EndpointName 'Moderation' -ApiBase $ApiBase
     }
 
     process {